requirements.txt
git+https://github.com/pyrogram/pyrogram@master
pytgcalls==3.0.0.dev16
python-dotenv==0.19.0
pytube==11.0.1
pafy==0.5.5
youtube-search==2.1.0
youtube-dl==2021.6.6
TgCrypto
